requirements.txt
streamlit>=1.28.0
openai>=1.0.0
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
feedparser>=6.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
pytz>=2023.3
lxml>=4.9.0
gspread>=6.1.2
google-auth>=2.25.2
python-dateutil>=2.9.0.post0
python-telegram-bot>=21.0
requests>=2.31.0
groq>=0.4.0
gspread>=6.1.2
google-auth>=2.25.2
python-dateutil>=2.9.0.post0
Pillow>=10.0.0
flask>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
pytz>=2023.3
lxml>=4.9.0
groq>=0.4.0
gspread>=6.1.2
google-auth>=2.25.2
python-dateutil>=2.9.0.post0